server/requirements.txt
connexion == 2.6.0
python_dateutil == 2.6.0
setuptools >= 21.0.0

mysqlclient~=2.0.3
Werkzeug~=1.0.1
six~=1.15.0
python-jose~=1.0.0
sendgrid~=6.6.0
Flask~=1.1.2